Added files for Section 5: JOINS
Challenge Problem 1. 
California sales tax laws have changed and we need to alert our customers to this through email.
What are the emails of the customers who live in California?

Challenge Problem 2. 
A customer walks in and is a huge fan of the actor "Nick Wahlberg" and wants to know which movies he is in.
Get a list of all the movies "Nick Wahlberg" has been in.